move runAsciidocInlineAction to AsciidocActionButtons.java
